+++ src/include/scatteractionsfinder.h
@@ -68,13 +68,13 @@ class ScatterActionsFinder : public ActionFinderInterface {
   /** Check the whole particle list for collisions
    * and return a list with the corrsponding Action objects. */
   ActionList find_actions_in_cell(const ParticleList &search_list,
-                                  float dt) const override;
+                                  double dt) const override;
   ActionList find_actions_with_neighbors(const ParticleList &search_list,
                                          const ParticleList &neighbors_list,
-                                         float dt) const override;
+                                         double dt) const override;
   ActionList find_actions_with_surrounding_particles(
       const ParticleList &search_list, const Particles &surrounding_list,
-      float dt) const override;
+      double dt) const override;
   /** Find some final collisions at the end of the simulation.
    * Currently does nothing. */
   ActionList find_final_actions(const Particles & /*search_list*/,
